requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv>=1.0.0
twilio